requirements.txt
nltk 
scikit-learn
numpy
pymorphy2
flask
requests
ffmpeg-python
